backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-multipart==0.0.6
pydantic==2.1.1
pydantic-settings==2.0.3
python-jose>=3.4.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.11.3
email-validator==2.0.0
python-dateutil==2.8.2
aiofiles==23.1.0

# OCR Dependencies
pytesseract==0.3.10
pdf2image==1.16.3
pillow==11.1.0